Se deja un solo SeleniumFunctions y un solo StepDefinitions
